requirements.txt
django
requests
django-contact-form
django-taggit
unidecode
Pillow
uwsgi
uwsgidecorators
pymysql
